Allow explicit specification of log file for the ats command
This allows running it on many files at once and then inspecting the log file to see how many files ended with a success and how many failed. I am not sure if this is a valid use case, or in other words, if this change is worth pulling into master. I am publishing it in case there is some interest in it.
Example of such usage:
find WIS-catalogue/ -exec pywmcp validate ats --log full-catalogue.log --file {} ;