Trial to generalize checking of keyword values (#7)
* Trial to generalize checking of keyword values

Added two examples of metadata files, one from Canada and one from Germany.
With the new code both validate successfully, originally the one from Germany failed on the test 9.1.1.

* Fixed E712

A code style "error".

* Fixed E302

Another code style error, sorry.

* Graceful handling of empty keyword element

* Decrease verbosity of a somewhat useless log entry

* Avoid duplicate running of the tests

I have left the call that is inside of a try/except block as this seems more useful but perhaps the first call was a feature that I am not getting now.
Tom, any comment to it?

* Revert "Avoid duplicate running of the tests"

This reverts commit 9e1197979e8ba9bb1bd64929c85a114391fcaa2b.
It should be on a separate branch to be able to make a spearate pull request for it

* Revert "Decrease verbosity of a somewhat useless log entry"

This reverts commit bfa4b5ad90181dd9396cdc1e746c7633eb041ff4.
It should be on a separate branch to be able to make a separate pull request for it. I _love_ GitHub.